switched bswap64 definition from compiler to OS
as suggested in #227,
because `clang` has a hard time properly inlining
either the scalar variant or __builtin_bswap64().

It doesn't matter much in the end,
as `bswap64()` is mostly useful for big endian systems,
and we don't know if `clang` inlining issues, visible on x64,
do expand to big-endian platforms.